Create separate file for view modifiers
